+++ packages/koi_api_bootstrap/lib/src/koi_api_log.dart
@@ -1,9 +1,8 @@
 enum KoiApiLogLevel { debug, info, warning, error, fatal }
 
-typedef KoiApiLogCallback =
-    void Function(
-      KoiApiLogLevel level,
-      String message,
-      Object? error,
-      StackTrace? stackTrace,
-    );
+typedef KoiApiLogCallback = void Function(
+  KoiApiLogLevel level,
+  String message,
+  Object? error,
+  StackTrace? stackTrace,
+);