+++ packages/koi_domain/lib/koi_domain.dart
@@ -1,2 +1,1 @@
 export 'entities/koi_user.dart';
-export 'entities/koi_workspace_tile.dart';